+++ semver-gradle-plugin/main/kotlin/com/javiersc/semver/gradle/plugin/internal/git/GitCache.kt
@@ -1,24 +1,61 @@
 package com.javiersc.semver.gradle.plugin.internal.git
 
 import com.javiersc.semver.Version
+import com.javiersc.semver.gradle.plugin.Commit
+import com.javiersc.semver.gradle.plugin.Tag
 import com.javiersc.semver.gradle.plugin.internal.InitialVersion
+import com.javiersc.semver.gradle.plugin.internal.semverWarningMessage
 import com.javiersc.semver.gradle.plugin.internal.warningLastVersionIsNotHigherVersion
+import com.javiersc.semver.gradle.plugin.services.hasNotCommits
+import java.io.File
 import org.eclipse.jgit.api.Git
 import org.eclipse.jgit.lib.Constants
 import org.eclipse.jgit.lib.ObjectId
 import org.eclipse.jgit.lib.Ref
 import org.eclipse.jgit.revwalk.RevCommit
 import org.eclipse.jgit.revwalk.RevWalk
+import org.eclipse.jgit.storage.file.FileRepositoryBuilder
+import org.gradle.api.provider.Provider
 
-internal class GitCache(private val git: Git) {
+private val checkedNotNullCache: GitCache
+    get() =
+        checkNotNull(gitCache) {
+            "`GitCache` must not be null at this point, report this as a bug, please"
+        }
+
+private var gitCache: GitCache? = null
+
+internal fun GitCache(
+    rootDir: File,
+    maxCount: Provider<Int>? = null,
+    refreshCache: Boolean
+): GitCache {
+    if (gitCache == null || refreshCache) {
+        gitCache = GitCache(rootDir, maxCount)
+    }
+    return checkedNotNullCache
+}
+
+internal class GitCache internal constructor(rootDir: File, maxCount: Provider<Int>? = null) {
+
+    private val gitDir: File by lazy { rootDir.resolve(".git") }
+
+    internal val git: Git by lazy {
+        Git(FileRepositoryBuilder().setGitDir(gitDir).readEnvironment().findGitDir().build()).also {
+            if (it.hasNotCommits()) {
+                semverWarningMessage("semver plugin can't work if there are no commits")
+            }
+        }
+    }
 
     internal val isClean: Boolean
         get() = git.status().call().isClean
 
-    internal val headRevCommit: RevCommit =
+    internal val headRevCommit: RevCommit by lazy {
         RevWalk(git.repository).parseCommit(git.repository.resolve(Constants.HEAD))
+    }
 
-    internal val headCommit: GitRef.Head =
+    internal val headCommit: GitRef.Head by lazy {
         GitRef.Head(
             GitRef.Commit(
                 message = headRevCommit.shortMessage,
@@ -26,10 +63,40 @@ internal class GitCache(private val git: Git) {
                 hash = headRevCommit.toObjectId().name,
             )
         )
+    }
 
-    internal val commitsInCurrentBranchRevCommit: List<RevCommit> = git.log().call().toList()
+    internal val commitsInCurrentBranchRevCommit: List<RevCommit> by lazy {
+        git.log().setMaxCount(maxCount?.get() ?: -1).call().toList()
+    }
+    internal val tagsInRepoRef: List<Ref> by lazy { git.tagList().call() }
 
-    internal val commitsInCurrentBranch: List<GitRef.Commit> =
+    internal val commitsInCurrentBranchHash: List<String> by lazy {
+        commitsInCurrentBranchRevCommit.map(RevCommit::getName)
+    }
+
+    internal val commitsInTheCurrentBranchPublicApi: List<Commit> by lazy {
+        commitsInCurrentBranchRevCommit.map { revCommit ->
+            val hash: String = revCommit.toObjectId().name
+            val tags: List<Tag> =
+                tagsInCurrentBranchRef
+                    .filter { ref -> commitHash(ref) == hash }
+                    .map { ref ->
+                        Tag(
+                            name = ref.tagName,
+                            refName = ref.name,
+                        )
+                    }
+            Commit(
+                message = revCommit.shortMessage,
+                fullMessage = revCommit.fullMessage,
+                hash = hash,
+                timestampEpochSecond = revCommit.authorIdent.whenAsInstant.epochSecond,
+                tags = tags,
+            )
+        }
+    }
+
+    internal val commitsInCurrentBranch: List<GitRef.Commit> by lazy {
         commitsInCurrentBranchRevCommit.map { revCommit ->
             revCommit.run {
                 GitRef.Commit(
@@ -39,11 +106,11 @@ internal class GitCache(private val git: Git) {
                 )
             }
         }
+    }
 
-    internal val lastCommitInCurrentBranch: GitRef.Commit? = commitsInCurrentBranch.firstOrNull()
-
-    internal val commitsInCurrentBranchHash: List<String> =
-        commitsInCurrentBranchRevCommit.map(RevCommit::getName)
+    internal val lastCommitInCurrentBranch: GitRef.Commit? by lazy {
+        commitsInCurrentBranch.firstOrNull()
+    }
 
     internal fun commitHash(ref: Ref): String = commitHash(ref.objectId)
 
@@ -55,8 +122,6 @@ internal class GitCache(private val git: Git) {
 
     internal val Ref.tagName: String
         get() = name.substringAfter("refs/tags/")
-
-    internal val tagsInRepoRef: List<Ref> = git.tagList().call()
 
     internal val tagsInCurrentBranchRef: List<Ref>
         get() = tagsInRepoRef.filter { ref -> commitHash(ref) in commitsInCurrentBranchHash }