+++ semver-gradle-plugin/main/kotlin/com/javiersc/semver/gradle/plugin/valuesources/VersionValueSource.kt
@@ -4,46 +4,38 @@ import com.javiersc.semver.Version
 import com.javiersc.semver.gradle.plugin.internal.calculatedVersion
 import com.javiersc.semver.gradle.plugin.internal.checkCleanProperty
 import com.javiersc.semver.gradle.plugin.internal.checkVersionIsHigherOrSame
+import com.javiersc.semver.gradle.plugin.internal.commitsMaxCount
 import com.javiersc.semver.gradle.plugin.internal.git.GitCache
 import com.javiersc.semver.gradle.plugin.internal.git.GitRef
 import com.javiersc.semver.gradle.plugin.internal.projectTagPrefix
 import com.javiersc.semver.gradle.plugin.internal.scopeProperty
 import com.javiersc.semver.gradle.plugin.internal.stageProperty
 import com.javiersc.semver.gradle.plugin.internal.tagPrefixProperty
-import com.javiersc.semver.gradle.plugin.services.GitBuildService
+import com.javiersc.semver.gradle.plugin.semverExtension
 import com.javiersc.semver.gradle.plugin.tasks.CreateSemverTagTask
 import com.javiersc.semver.gradle.plugin.tasks.PushSemverTagTask
-import org.eclipse.jgit.api.Git
-import org.eclipse.jgit.storage.file.FileRepositoryBuilder
 import org.gradle.api.Project
-import org.gradle.api.file.DirectoryProperty
-import org.gradle.api.provider.ListProperty
+import org.gradle.api.file.RegularFileProperty
 import org.gradle.api.provider.Property
 import org.gradle.api.provider.Provider
 import org.gradle.api.provider.ValueSource
 import org.gradle.api.provider.ValueSourceParameters
 import org.gradle.kotlin.dsl.of
 
-public abstract class VersionValueSource : ValueSource<String, VersionValueSource.Params> {
-
-    private val cache =
-        GitCache(
-            Git(
-                FileRepositoryBuilder()
-                    .setGitDir(parameters.gitDir.get().asFile)
-                    .readEnvironment()
-                    .findGitDir()
-                    .build(),
-            ),
-        )
+internal abstract class VersionValueSource : ValueSource<String, VersionValueSource.Params> {
 
     override fun obtain(): String =
         with(parameters) {
-            val isSamePrefix = tagPrefixProperty.get() == projectTagPrefix.get()
+            val isSamePrefix: Boolean = tagPrefixProperty.get() == projectTagPrefix.get()
 
-            val lastSemver = cache.lastVersionInCurrentBranch(projectTagPrefix.get())
+            val cache = GitCache(parameters.rootDir.get().asFile, parameters.commitsMaxCount)
+
+            val lastSemver: Version = cache.lastVersionInCurrentBranch(projectTagPrefix.get())
             val lastVersionInCurrentBranch =
                 cache.versionsInCurrentBranch(projectTagPrefix.get()).map(Version::toString)
+
+            val lastVersionCommitInCurrentBranch =
+                cache.lastVersionCommitInCurrentBranch(projectTagPrefix.get())?.hash
 
             val version: String =
                 calculatedVersion(
@@ -59,9 +51,9 @@ public abstract class VersionValueSource : ValueSource<String, VersionValueSourc
                     clean = cache.isClean,
                     checkClean = checkClean.get(),
                     lastCommitInCurrentBranch = cache.lastCommitInCurrentBranch?.hash,
-                    commitsInCurrentBranch = commitsInCurrentBranch.get(),
-                    headCommit = headCommit.get(),
-                    lastVersionCommitInCurrentBranch = lastVersionCommitInCurrentBranch.orNull,
+                    commitsInCurrentBranch = cache.commitsInCurrentBranch.map(GitRef.Commit::hash),
+                    headCommit = cache.headCommit.commit.hash,
+                    lastVersionCommitInCurrentBranch = lastVersionCommitInCurrentBranch,
                 )
 
             checkVersionIsHigherOrSame(version, lastSemver)
@@ -70,49 +62,30 @@ public abstract class VersionValueSource : ValueSource<String, VersionValueSourc
         }
 
     internal interface Params : ValueSourceParameters {
-        val gitDir: DirectoryProperty
+        val rootDir: RegularFileProperty
+        val commitsMaxCount: Property<Int>
         val tagPrefixProperty: Property<String>
         val projectTagPrefix: Property<String>
         val stageProperty: Property<String?>
         val scopeProperty: Property<String?>
         val creatingSemverTag: Property<Boolean>
         val checkClean: Property<Boolean>
-        val commitsInCurrentBranch: ListProperty<String>
-        val headCommit: Property<String>
-        val lastVersionCommitInCurrentBranch: Property<String?>
     }
 
-    public companion object {
+    companion object {
 
-        internal fun register(
-            project: Project,
-            gitTagBuildService: Provider<GitBuildService>,
-        ): Provider<String> =
+        fun register(project: Project): Provider<String> =
             project.providers.of(VersionValueSource::class) { valueSourceSpec ->
-                val cache = gitTagBuildService.map(GitBuildService::gitCache).get()
-
-                valueSourceSpec.parameters.gitDir.set(
-                    project.objects
-                        .directoryProperty()
-                        .fileProvider(
-                            project.provider {
-                                project.file("${project.rootProject.projectDir}/.git")
-                            },
-                        ),
-                )
+                valueSourceSpec.parameters.rootDir.set(project.semverExtension.rootDir)
+                val commitsMaxCount: Int =
+                    project.commitsMaxCount.orNull ?: project.semverExtension.commitsMaxCount.get()
+                valueSourceSpec.parameters.commitsMaxCount.set(commitsMaxCount)
                 valueSourceSpec.parameters.projectTagPrefix.set(project.projectTagPrefix.get())
                 valueSourceSpec.parameters.tagPrefixProperty.set(project.tagPrefixProperty.get())
                 valueSourceSpec.parameters.stageProperty.set(project.stageProperty.orNull)
                 valueSourceSpec.parameters.scopeProperty.set(project.scopeProperty.orNull)
                 valueSourceSpec.parameters.creatingSemverTag.set(project.isCreatingSemverTag)
                 valueSourceSpec.parameters.checkClean.set(project.checkCleanProperty.get())
-                valueSourceSpec.parameters.commitsInCurrentBranch.set(
-                    cache.commitsInCurrentBranch.map(GitRef.Commit::hash),
-                )
-                valueSourceSpec.parameters.headCommit.set(cache.headCommit.commit.hash)
-                valueSourceSpec.parameters.lastVersionCommitInCurrentBranch.set(
-                    cache.lastVersionCommitInCurrentBranch(project.projectTagPrefix.get())?.hash,
-                )
             }
     }
 }
@@ -120,5 +93,5 @@ public abstract class VersionValueSource : ValueSource<String, VersionValueSourc
 private val Project.isCreatingSemverTag: Boolean
     get() =
         gradle.startParameter.taskNames.any { taskName: String ->
-            taskName == CreateSemverTagTask.taskName || taskName == PushSemverTagTask.taskName
+            taskName == CreateSemverTagTask.TaskName || taskName == PushSemverTagTask.TaskName
         }