+++ semver-gradle-plugin/testFixtures/kotlin/com/javiersc/semver/gradle/plugin/Sandbox.kt
@@ -4,6 +4,7 @@ import java.io.File
 import java.nio.file.Files
 import java.nio.file.Path
 import java.nio.file.Paths
+import java.time.Instant
 import org.eclipse.jgit.api.Git
 import org.eclipse.jgit.storage.file.FileRepositoryBuilder
 import org.gradle.testkit.runner.GradleRunner
@@ -40,7 +41,7 @@ internal fun createSandboxFile(prefix: String): File =
 internal fun File.generateInitialCommitAddVersionTagAndAddNewCommit(
     doBefore: Git.() -> Unit = {},
     doAfter: Git.() -> Unit = {},
-) {
+): List<Commit> {
     createGitIgnore()
     val git: Git = Git.init().setDirectory(this).call()
     doBefore(git)
@@ -57,6 +58,33 @@ internal fun File.generateInitialCommitAddVersionTagAndAddNewCommit(
     git.add().addFilepattern(".").call()
     git.commit().setMessage("Add new2").call()
     doAfter(git)
+
+    return listOf(
+        Commit(
+            message = "Add new2",
+            fullMessage = "Add new2",
+            hash = "f099fed42808c387d38f8e34934d4ef2d6f3c2c5",
+            timestampEpochSecond = Instant.now().epochSecond,
+            tags = emptyList()
+        ),
+        Commit(
+            message = "Add new",
+            fullMessage = "Add new",
+            hash = "0d9be4d12a836435c91fc38e0d64bcd3c14c89c8",
+            timestampEpochSecond = Instant.now().epochSecond,
+            tags = emptyList()
+        ),
+        Commit(
+            message = "Initial commit",
+            fullMessage = "Initial commit",
+            hash = "e2592aafc259bd797e32f1ea4fe60cc7c0698f70",
+            timestampEpochSecond = Instant.now().epochSecond,
+            tags =
+                listOf(
+                    Tag(name = "1.0.0", refName = "refs/tags/1.0.0"),
+                )
+        )
+    )
 }
 
 internal fun File.createGitIgnore() {